+++ packages/client/src/pages/main/modules/leaderBoard/components/LeaderBoardAvatar/constants.ts
